% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{add_question_detail}
\alias{add_question_detail}
\title{Add question type, clean text, notes, skip logic to question elements}
\usage{
add_question_detail(questions, blocks, qtNotesList)
}
\arguments{
\item{questions}{A list of questions extracted from a Qualtrics survey file. See
`valid_questions_blocks_from_survey` to create question list.}

\item{blocks}{A list of blocks extracted from a Qualtrics Survey file. See
`valid_questions_blocks_from_survey` to generate this list of blocks.}

\item{qtNotes}{A list of user notes extracted from a Qualtrics survey file. Use
`note_text_from_survey` to create the user note list. Elements of this list
are named with the parent QID.}
}
\value{
A list of survey questions with added elements for Question Text cleaned of
HTML and CSS, human readable question type, user notes, and skip logic.
}
\description{
Given a list of questions, blocks, and notes from a survey, this function
will add information to the question elements that will be used in later
QualtricsTools functions.
}
\details{
Information added to these questions are elements within the question list.
The earlier version of QualtricsTools added these to the Payload, but it is more easily
to access them as named items directly within the quesiton. They are not part of the
Payload from Qualtrics, so it seems fine to pull the information from here.
}
