% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{app}
\alias{app}
\title{A Shiny app to format Qualtrics survey data and generate reports}
\usage{
qualtrics::app()
}
\description{
This function launches the Shiny interface for the Qualtrics
package from the files in the install or 'inst' directory.
}

