% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{ask_for_csv}
\alias{ask_for_csv}
\title{Ask the user for the Qualtrics Response Set}
\usage{
ask_for_csv(responsesfile, headerrows)
}
\arguments{
\item{responsesfile}{this is the file path of a CSV response set to a Qualtrics survey}

\item{headerrows}{the number of rows before responses begin in the CSV data}
}
\value{
a list of two elements: the responses data frame, and the original_first_rows data frame
}
\description{
ask_for_csv() inputs a file and a number of headerrows and returns a
list containing a responses data frame, and a 1-2 row data frame of the
responses' headers. Both input parameters are optional in the sense that
they do not have to be provided inline to the function, and that the
function will ask the user to choose a file for the responsesfile if one is
not provided inline, and it will default to using 3 headerrows.
}
