% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{ask_user_for_csv}
\alias{ask_user_for_csv}
\title{Ask the user for the Qualtrics Response Set}
\usage{
ask_user_for_csv(responsesfile)
}
\value{
The csv file the user uploads, as a data frame
}
\description{
This function automatically turns the entries in the
first row into a "text" attribute of the column. Additionally,
it removes the first row after turning its entries into attributes.
}

