% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{blocks_from_survey}
\alias{blocks_from_survey}
\title{Generate a List of Survey Blocks}
\usage{
blocks_from_survey(survey)
}
\arguments{
\item{survey}{This should be a Qualtrics survey in the form of a list
imported from the JSON-formatted QSF file.}
}
\value{
The blocks element returned is a list of blocks containing for
each a type, description, ID, and BlockElements (which contains the
list of questions included in a given block).
}
\description{
This function takes the survey blocks out of the survey
and turns them into their own list for later use.
Additionally, the block element can be further reduced to include
only the payload of the survey blocks.
}

