% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{choice_text_from_response_column}
\alias{choice_text_from_response_column}
\title{Get the Choice Text from the First Row of the Responses}
\usage{
choice_text_from_response_column(response_column, original_first_row, blocks)
}
\arguments{
\item{response_column}{The name of a response column from the response set}

\item{original_first_row}{The first row of the original response set. If you have
the original_first_rows, you can use original_first_rows[1,].}

\item{blocks}{A list of the survey blocks, with the questions included in them}
}
\value{
The choice text corresponding to a response column
}
\description{
This function uses the first row of the response data from Qualtrics
to determine the choice text a response column corresponds to.
}

