% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{clean_question_text}
\alias{clean_question_text}
\title{Clean Question Text}
\usage{
clean_question_text(question)
}
\arguments{
\item{question}{A survey question with Payload element}
}
\value{
The question with QuestionTextClean added to the Payload
}
\description{
Clean Question Text from the QuestionText Element of Payload and return a question
with added cleanQuestionText
}
\details{
Given a question that includes a Payload element with QuestionText, clean the Question Text
of HTML and CSS and return the question with element QuestionTextClean added to the payload.
}
