% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{create_merged_response_column}
\alias{create_merged_response_column}
\title{Add a Custom Response Column with Merged Contents}
\usage{
create_merged_response_column(response_columns, col_name, question_blocks,
  survey_responses)
}
\arguments{
\item{response_columns}{a list of names of response columns}

\item{col_name}{the desired name for the added output column}

\item{question_blocks}{a list of blocks with questions inserted
in place of the BlockElements.}

\item{survey_responses}{the data frame of responses as imported by
ask_for_csv()}
}
\value{
a response data frame with a single new column which
represents the contents of the specified response_columns.
}
\description{
The create_merged_response_column() function allows a user to
create a response set with a column reflecting the combined
contents of several different columns. If the user specifies
response columns that are part of the survey question responses,
then the responses are coded and the coded values are used
instead of the variable response values. The column is appended to
the response set and the response set with the additional
new column is returned to the user.
}
