% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{create_panel_data}
\alias{create_panel_data}
\title{Create Panel Data for Reshaped Data}
\usage{
create_panel_data(panel_columns, survey_responses, lean_responses,
  question_dict)
}
\arguments{
\item{panel_columns}{a list of names of columns desired for inclusion}

\item{survey_responses}{a response set data frame imported from the Qualtrics CSV responses}

\item{lean_responses}{an optionally included data frame generated by the lean_responses() function}

\item{question_dict}{an optionally included data frame generated by the create_response_column_dictionary() function}
}
\description{
The user provides this function a character vector of names of response columns (
as the panel_columns var), along with the response set data frame, and the returned
data is a data frame with respondent IDs and column(s) for each column specified.
If the optional parameters, lean_responses and question_dict are provided, then
the function will use them to include a "Raw Response" and a "Coded Response"
column for any columns which are question response columns.
}

