% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{create_question_dictionary_from_qsf}
\alias{create_question_dictionary_from_qsf}
\title{Create a response column dictionary with only the qsf}
\usage{
create_question_dictionary_from_qsf(qsf_path)
}
\description{
The user provides the survey qsf, and this creates a response column dictionary
The data .csv is not required
This can be used to identify quesiton types in the original survey, regardless
of whether you changed questions for a shell survey.
}
