% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{create_response_column_dictionary}
\alias{create_response_column_dictionary}
\title{Create a Response Column Dictionary}
\usage{
create_response_column_dictionary(question_blocks, orig_first_row)
}
\arguments{
\item{question_blocks}{use questions_into_blocks() to create a list of blocks with
the survey's questions inserted appropriately into them.}

\item{orig_first_row}{this is the first row of the response data CSV, it is
automatically provided to you when you use get_setup() or in the shiny application.}
}
\value{
a dataframe detailing in each row the response columns and their description.
}
\description{
By providing the questions structured by their blocks and the original first
row of the response CSV from Qualtrics, this function is able to create a
dictionary where each row is an entry for a response column. The response
columns are listed along with the data export tag of the question they correspond
to, their question stem and question choice, the question types (1, 2, and 3 levels
of question types!), and the response type.
}

