% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{create_response_lookup_table}
\alias{create_response_lookup_table}
\title{Create a Lookup Table for Response Variables}
\usage{
create_response_lookup_table(question)
}
\arguments{
\item{question}{A list of questions extracted from a Qualtrics QSF file with
responses inserted into them.}
}
\value{
A dataframe with columns "var", "recode_value", and "text", where each
row corresponds to a unique variable response to the question.
}
\description{
For each variable response in the responses to a question,
this function creates a row which contains the variable,
its recode value if the question uses recode values, and the
corresponding text for the given variable response. The exact
format of the output lookup table (dataframe) is dictated
by the question type depending on whether it is a
multiple answer question, a multiple choice single answer question,
or a matrix single answer question. Note that this function
only creates a lookup table for each response which appears in the responses.
}
