% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{display_logic_from_question}
\alias{display_logic_from_question}
\title{Return a list of a Question's Display Logic Components}
\usage{
display_logic_from_question(question)
}
\arguments{
\item{question}{A qualtrics survey question}
}
\value{
an ordered list of display logic messages
}
\description{
For each question, if they appear, go through the
Question's DisplayLogic, each Choice's DisplayLogic, and
each Answer's DisplayLogic. For each of them, use clean_html
to format them, and then add them to the list. If a question
has any of these display logic components, insert before
adding any display logic a line detailing what part of the
question the following display logic corresponds to.
}

