% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_question_index_by_qid}
\alias{find_question_index_by_qid}
\title{Find a Question by its QuestionID}
\usage{
find_question_index_by_qid(questions, qid)
}
\description{
Find a Question by its QuestionID
}
