% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{flow_from_survey}
\alias{flow_from_survey}
\title{Get the Flow out of the Survey}
\usage{
flow_from_survey(survey)
}
\description{
The 'Flow' is a list of Block IDs in the order that they are presented
in the survey as it is taken by a respondent. The flow list that is returned
from this function is used by functions like text_appendices_table and
tabelize_blocks to get the ordering of the survey in the preview correct.
}

