% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{format_and_split_comment_sheets}
\alias{format_and_split_comment_sheets}
\title{Format and Split a list of Unprocessed Coded Comment Sheets}
\usage{
format_and_split_comment_sheets(
  coded_comment_sheets,
  responses,
  split_column,
  code_type
)
}
\arguments{
\item{coded_comment_sheets}{A list of dataframes for
each sheet of coded comments.}

\item{responses}{A dataframe of Qualtrics responses to a survey.
Use ask_for_csv() to create such a dataframe from a CSV file.}

\item{split_column}{The string name of the column across which the coded comments
should be split.}

\item{code_type}{type of coded comment data export. The default is
to use NVivo crosstab export with the ResponseID in the first row and
second column labeled with the varnmae and containing response presence
1/0 indicator. For old filemaker pro format, use "fmp" specification.
(\code{"nvivo"} or \code{"fmp"})}
}
\value{
A list of lists, for each split group, and for the pairs of question IDs with
coded comment tables in each split group.
}
\description{
When splitting the respondents of a survey to create split reports,
the coded comments are split by this function and then returned as a list of lists.
The first list is a list for each split group, and each list within those is a
list of pairs of question IDs and their coded comments tables.
}
