% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{format_coded_comment_sheets}
\alias{format_coded_comment_sheets}
\title{Format Comment Coding Data into Frequency Tables}
\usage{
format_coded_comment_sheets(coded_comment_sheets, code_type)
}
\arguments{
\item{coded_comment_sheets}{A list of dataframes for
each sheet of coded comments.}

\item{code_type}{type of coded comment data export. The default is
to use NVivo crosstab export with the ResponseID in the first row and
second column labeled with the varnmae and containing response presence
1/0 indicator. For old filemaker pro format, use "fmp" specification.
(\code{"nvivo"} or \code{"fmp"})}
}
\value{
A list of pairs (varname, coded_table) where varname corresponds
to the response column name of the comments coded and coded_table
summarizes the frequencies of the provided coded comments.
}
\description{
This function takes as an argument a list of dataframes
which categorize the text responses to a free-response
component of a survey and formats them into a standardized
frequency table by applying format_coded_comments to each
coded_comment_sheet.
}
