% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{generate_coded_comments}
\alias{generate_coded_comments}
\title{Create Text Appendices including Coded Comments}
\usage{
generate_coded_comments(qsf_path, csv_path, sheets_dir, output_dir,
  filename = "text appendices.docx", n_threshold = 16, headerrows)
}
\description{
Using `get_setup`, `directory_get_coded_comment_sheets`, `format_coded_comment_sheets`,
`insert_coded_comments`, and `html_2_pandoc`, this function renders
text appendices with coded comments included from CSV or XLSX files
from the specified `sheets_dir` parameter.
}

