% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{generate_results}
\alias{generate_results}
\title{Create Results Tables and Pair Them to Questions}
\usage{
generate_results(questions, original_first_rows)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics Survey File containing
response columns (inserted by link_responses_to_questions).}

\item{original_first_rows}{A dataframe contianing the header information
for each column of response data. This dataframe includes a row for the DataExportTag based
response column names, another for the Question Text stem and choice text (although
truncated), and a row with QID based column names.}
}
\value{
A list of questions with their results tables paired to them
under the questions[[i]][['Table']]
}
\description{
The generate_results function takes a list of questions which have
their responses paired to them, determines their question type,
uses the results generation functions to create their results table,
and saves the table to the question's [['Table']] element. The function
returns the list of questions with their paired results tables.
}
