% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{generate_split_coded_comments}
\alias{generate_split_coded_comments}
\title{Split a Survey's Split Coded Comment Appendices}
\usage{
generate_split_coded_comments(qsf_path, csv_path, sheets_dir, output_dir,
  split_by, n_threshold = 15, headerrows)
}
\description{
This question automates the entire process of splitting a
survey's text appendices by specific response columns. The QSF
and CSV file are passed as string arguments,
the sheets_dir specifies where the coded comments excel or csv
data is stored, and the output_dir specifies where the split
coded comment appendices should be saved. The n_threshold
specifies how many coded comments there must be before the coded
comment appendices are included, and headerrows is an argument
necessary to process the survey results correctly.
}

