% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding_NVivo.R
\name{get_coded_comment_sheet_NVivo}
\alias{get_coded_comment_sheet_NVivo}
\title{Turn a Single Coded File into a Data Frame}
\usage{
get_coded_comment_sheet_NVivo(codedfile)
}
\arguments{
\item{codedfile}{The string path to a Excel file.}
}
\value{
A dataframe version of the contents of the coded comments
in the codedfile.
}
\description{
This retrieves comment coding data as a dataframe
from a Excel file. Previously with FileMaker exports, the sheet was
identified by the sheet name "coded" (case did not matter).
With the NVivo exports, there should only be one sheet. We have decided
to add a check that the sheet must be called "coded" when there is more than
one sheet to choose from.
}
