% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{get_coded_questions_and_blocks}
\alias{get_coded_questions_and_blocks}
\title{Get Questions (with Responses) and Blocks (with Questions)}
\usage{
get_coded_questions_and_blocks(survey, responses)
}
\arguments{
\item{survey}{A qualtrics survey, uploaded from a qsf/json file}

\item{responses}{Qualtrics responses to a survey, uploaded from csv as a data frame}
}
\value{
a list with two elements, the first being the survey questions,
and the second being the survey blocks
}
\description{
This function returns a list with two elements, where
the first element is questions with their responses
and results-tables listed in them, and secondly
the blocks of the survey, with questions listed in
each block as BlockElements.
}

