% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_setup_in_environment}
\alias{get_setup_in_environment}
\title{Load Survey Data into an Arbitrary Environment in R}
\usage{
get_setup_in_environment(qsf_path, csv_path, headerrows, environment)
}
\arguments{
\item{qsf_path}{The string location of the survey as a .QSF (Qualtrics Survey File)}

\item{csv_path}{The string location of the survey's responses, downloaded from Qualtrics}

\item{headerrows}{An optional parameter for specifying the number of
headerrows in the response csv.}

\item{environment}{An R environment to save each of the survey, responses, questions, blocks
original_first_rows, and flow into.}
}
\description{
This function puts the survey, responses, questions, blocks
original_first_rows, and flow into the environment specified in R.
qsf_path, csv_path, and headerrows are optional. If qsf_path and
csv_path are provided, then the function uses get_setup with return_data_as_list
to process the survey data and then insert the returned data into the
specified environment. If the qsf_path and csv_path are not specified,
the function first checks the global scope to see if all output from
get_setup exists, and if so, the function copies them into the specified
environment. If the qsf_path and csv_path are not provided, and the
output of get_setup is not in the global scope, then the function calls
get_setup to interactively ask the user to select the qsf_path and csv_path.
Headerrows is defaulted to 3 by the get_setup function if it is not provided.
}
