% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{has_na}
\alias{has_na}
\title{Determine if a question has an NA type choice by looking for negative recode values.
Using negative recode values to indicate N/A
type choices is the convention used in this project to identify N/A type
choices regardless of specific question text. N/A type choices may be "Not Applicable"
or some other option that is tabled separately from the other, "valid" responses.}
\usage{
has_na(question)
}
\arguments{
\item{question}{The question parameter is a single question from a qualtrics survey.}
}
\value{
The return value of this is a boolean, true if it has any NA type choices
indicated by negative recode values and false otherwise.
}
\description{
Determine if a question has an NA type choice by looking for negative recode values.
Using negative recode values to indicate N/A
type choices is the convention used in this project to identify N/A type
choices regardless of specific question text. N/A type choices may be "Not Applicable"
or some other option that is tabled separately from the other, "valid" responses.
}
