% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{has_single_te_component}
\alias{has_single_te_component}
\title{Determine if a question has a single text entry component}
\usage{
has_single_te_component(question)
}
\arguments{
\item{question}{The question parameter is a single question from a qualtrics survey.}
}
\value{
The return value of this is a boolean, TRUE if there is a single text entry
component and FALSE otherwise.
}
\description{
A question has a single text entry component if it is not a Text Entry Question
(QuestionType != "TE") and the text string "TEXT" appears in the name of one response column.
}
