% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc_conversion.R
\name{html_2_pandoc}
\alias{html_2_pandoc}
\title{Convert HTML to a Docx File}
\usage{
html_2_pandoc(html, file_name, format, output_dir)
}
\arguments{
\item{html}{an html body contents string. Do not include the <html><body> ... tags,
the html parameter will automatically have those kinds of tags added within this
function.}

\item{output_dir}{an optional parameter for specifying an output folder}
}
\value{
the path to the docx file created from converting the html.
}
\description{
This function uses pandoc and tempfiles to create a docx version
of an HTML document. The HTML is prepended and appended with
<!doctype html><html><body> and </body></html>. Then it is saved
to a tempfile, and then pandoc converts it to a .docx file. Finally,
the path to the .docx file is returned.
}
