% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_rendering.R
\name{html_to_docx}
\alias{html_to_docx}
\title{Convert HTML to a Docx File}
\usage{
html_to_docx(html)
}
\arguments{
\item{html}{an html body contents string. Do not include the <html><body> ... tags,
the html parameter will automatically have those kinds of tags added within this
function.}
}
\value{
the path to the docx file created from converting the html.
}
\description{
This function uses pandoc and tempfiles to create a docx version
of an HTML document. The HTML is prepended and appended with
<!doctype html><html><body> and </body></html>. Then it is saved
to a tempfile, and then pandoc converts it to a .docx file. Finally,
the path to the .docx file is returned.
}

