% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{human_readable_qtype}
\alias{human_readable_qtype}
\title{Create Human Readable Question Types}
\usage{
human_readable_qtype(questions)
}
\arguments{
\item{questions}{A list of questions extracted from a Qualtrics QSF file. Use
questions_from_survey() to get them from an imported survey.}
}
\value{
A list of questions which include in their Payload a QuestionTypeHuman field.
}
\description{
This function saves, admittedly reductionist, more friendly question type
descriptions. It doesn't have human human readable versions for every question type,
but for some it is useful to reduce the question type to something more simple.
This has a nested function (create_qtype) that determines a question's human question type,
and then it loops through the questions and saves to each the QuestionTypeHuman
field with the output of create_qtype.
}
