% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{insert_coded_comments}
\alias{insert_coded_comments}
\title{Insert Coded Comments into Blocks}
\usage{
insert_coded_comments(blocks, original_first_rows, coded_comments)
}
\description{
Insert Coded Comments into Blocks
}

