% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{insert_notes_into_questions}
\alias{insert_notes_into_questions}
\title{Insert the Notes for a question into its qtNotes}
\usage{
insert_notes_into_questions(questions, notes)
}
\description{
Insert the Notes for a question into its qtNotes
}
