% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{insert_questions_into_block}
\alias{insert_questions_into_block}
\title{Insert questions into blocks}
\usage{
insert_questions_into_block(block, questions)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics survey}

\item{blocks}{A list of blocks from a Qualtrics survey}
}
\value{
The list of blocks with question elements updated to include the full question.
}
\description{
Insert questions into blocks
}
