% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{insert_questions_into_blocks}
\alias{insert_questions_into_blocks}
\title{Iterate over blocks and insert questions to each}
\usage{
insert_questions_into_blocks(questions, blocks)
}
\arguments{
\item{questions}{A list of questions}

\item{blocks}{A list of survey blocks}
}
\value{
A list of blocks with the questions inserted into the appropriate block elements
}
\description{
This iterates over a list of blocks and inserts questions to each of the survey blocks.
}
