% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{insert_skiplogic_into_questions}
\alias{insert_skiplogic_into_questions}
\title{Insert skip logic into questions}
\usage{
insert_skiplogic_into_questions(questions, blocks)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics survey}

\item{blocks}{A list of blocks from a Qualtrics survey}
}
\value{
A list of questions with skip logic added to questions where appropriate
}
\description{
Insert skip logic into questions
}
