% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_matrix_multiple_answer}
\alias{is_matrix_multiple_answer}
\title{Determine if a question is a matrix and multiple answer question}
\usage{
is_matrix_multiple_answer(question)
}
\arguments{
\item{question}{The question parameter is a single question from a qualtrics survey.}
}
\value{
The return value of this is a boolean, true if it is one of these kinds of
questions and false otherwise.
}
\description{
A question is considered a matrix multiple answer question if it
is a `[['Payload']][['QuestionType']] == "Matrix"` question with
`[['Payload']][['SubSelector']] == "MultipleAnswer"`
}

