% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_functions.R
\name{is_matrix_question}
\alias{is_matrix_question}
\title{Determine if a question is a Matrix question}
\usage{
is_matrix_question(x)
}
\value{
The return value of this is a boolean, true if it is one of these kinds of
questions and false otherwise.
}
\description{
Determine if a question is a Matrix question
}

