% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_mc_single_answer}
\alias{is_mc_single_answer}
\title{Determine if a question is a single answer question}
\usage{
is_mc_single_answer(question)
}
\arguments{
\item{question}{The question parameter is a single question from a qualtrics survey.}
}
\value{
The return value of this is a boolean, true if it is one of these kinds of
questions and false otherwise.
}
\description{
Each of the is-functions defined in the qualtrics package are used
for determining which response parsing function should be used. A
function is considered multiple choice if it is listed in
the qsf file as having [['Payload']][['QuestionType']] == "MC"
(standing for Multiple Choice), and the `[['Payload']][['Selector']]` is set to one of the
following:
"Single Answer Vertical",
"Single Answer Horizontal",
"Single Answer Column",
"Dropdown List",
"Select Box",
}
