% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_multiple_answer}
\alias{is_multiple_answer}
\title{Determine if a question is a multiple answer question}
\usage{
is_multiple_answer(x)
}
\description{
Determine if a question is a multiple answer question
}
