% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_functions.R
\name{is_multiple_choice}
\alias{is_multiple_choice}
\title{Determine if a question is a multiple choice question}
\usage{
is_multiple_choice(question)
}
\arguments{
\item{question}{The question parameter is a single question from a qualtrics survey.}
}
\value{
The return value of this is a boolean, true if it is one of these kinds of
questions and false otherwise.
}
\description{
Each of the is-functions defined in the qualtrics package are used
for determining which response parsing function should be used. A 
function is considered multiple choice if it is listed in 
the qsf file as having $Payload$QuestionType == "MC" 
(standing for Multiple Choice), and the `$Payload$Selector` is set to one of the 
following:
"Multiple Answer Vertical", 
"Multiple Answer Horizontal",
"Multiple Choice Select Box",
"Multiple Answer Column", 
or if it is a `$Payload$QuestionType == "Matrix"` question with 
`$Payload$SubSelector == "MultipleAnswer"`
}

