% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_functions.R
\name{is_rank_order}
\alias{is_rank_order}
\title{Determine if a question is a rank order question}
\usage{
is_rank_order(x)
}
\value{
The return value of this is a boolean, true if it is one of these kinds of
questions and false otherwise.
}
\description{
Determine if a question is a rank order question
}

