% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_single_answer}
\alias{is_single_answer}
\title{Determine if a question is a single answer question}
\usage{
is_single_answer(x)
}
\description{
Determine if a question is a single answer question
}

