% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_TE_numerical}
\alias{is_TE_numerical}
\title{Determine if a question is a text entry question with a numerical validation}
\usage{
is_TE_numerical(question)
}
\description{
Determine if a question is a text entry question with a numerical validation
}
