% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_text_entry_numeric}
\alias{is_text_entry_numeric}
\title{Determine if a question is a text entry question with a numerical
validation}
\usage{
is_text_entry_numeric(question)
}
\arguments{
\item{question}{The question parameter is a single question from a qualtrics survey.}
}
\value{
The return value of this is a boolean, true if it is one of these kinds of
questions and false otherwise.
}
\description{
A question is considered a text entry with numeric validation if it
is a `[['Payload']][['QuestionType']] == "TE"` question with
`[['Payload']][['Validation']][['Settings']][['Type']] == "ContentType"
and `[['Payload']][['Validation']][['Settings']][['ContentType']] ==
"ValidNumber"
}
