% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{lean_responses}
\alias{lean_responses}
\title{Create Long and Lean Response Dictionary}
\usage{
lean_responses(panel_columns)
}
\arguments{
\item{panel_columns}{A list of names of response columns to include in the
output formatted as panel data.}
}
\value{
a data frame with each row detailing an individual survey response.
}
\description{
lean_responses() creates a data frame where each row corresponds to
an individual response to the survey. Each response contains
the respondents' id, the question data export tag, the response column's name,
the question text, the question type(s), the variable response, and the coded
response, and then that individual's responses to any questions chosen for inclusion
as panel data.
}

