% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{load_csv_data}
\alias{load_csv_data}
\title{Set Response Data to Sample Data or User Data}
\usage{
load_csv_data(file2, file1, headerrows)
}
\arguments{
\item{file1}{This should be a CSV received from a file upload made in the Shiny UI,
which includes a file1[['datapath']] where the data can be located.}
}
\value{
The return value is the responses data frame
}
\description{
load_csv_data returns the sample response set or the user's response set depending
on whether or not the user has uploaded data.
}
