% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{load_qsf_data}
\alias{load_qsf_data}
\title{Set Survey to Sample Survey or User Survey}
\usage{
load_qsf_data(file2)
}
\arguments{
\item{file2}{This should be a QSF file received from a file upload made in the Shiny UI,
which includes a file1$datapath where the data can be located.}
}
\value{
The return value is the survey list object
}
\description{
load_qsf_data returns the sample survey or the user's survey depending
on whether or not the user has uploaded data.
}

