% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_coded_comments}
\alias{make_coded_comments}
\title{Create Text Appendices including Coded Comments}
\usage{
make_coded_comments(qsf_path, csv_path, headerrows, sheets_dir, output_dir,
  filename = "text appendices.docx", n_threshold = 15)
}
\arguments{
\item{qsf_path}{(optional) is the string path location of the .qsf file to be processed.}

\item{csv_path}{(optional) is the string path location of the .csv file to be processed.}

\item{headerrows}{(optional) specifies the number of header rows in the CSV data.}

\item{sheets_dir}{is the string path location of the directory which contains Excel documents
with a "Coded" sheet formatted as specified on the wiki:
https://github.com/ctesta01/QualtricsTools/wiki/Comment-Coding}

\item{output_dir}{specifies the path of the directory to save the output file in.}

\item{filename}{specifies the name of the output file.}

\item{n_threshold}{is the number of verbatim comments which will appear before being truncated.}
}
\description{
Using `get_setup`, `directory_get_coded_comment_sheets`, `format_coded_comment_sheets`,
`insert_coded_comments`, and `html_2_pandoc`, this function renders
text appendices with coded comments included from CSV or XLSX files
from the specified `sheets_dir` parameter.
}
