% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_results_tables}
\alias{make_results_tables}
\title{Export a file containing the results tables}
\usage{
make_results_tables(qsf_path, csv_path, headerrows, output_dir,
  filename = "Results Tables.docx")
}
\arguments{
\item{qsf_path}{(optional) is the string path location of the .qsf file to be processed.}

\item{csv_path}{(optional) is the string path location of the .csv file to be processed.}

\item{headerrows}{(optional) specifies the number of header rows in the CSV data.}

\item{output_dir}{specifies the path of the directory to save the output file in.}

\item{filename}{specifies the name of the output file.}
}
\description{
`make_results_tables` uses `get_setup` and `html_2_pandoc` to process a
survey and then save its results into a file. If the `qsf_path,` and `csv_path`
are included as parameters, then they will be passed to `get_setup` along with a
`return_data=TRUE` parameter in order to return the survey, responses,
questions, blocks, original_first_rows, and flow as variables local to the function
scope. If they are not passed, they will be retrieved as needed from the global scope.
The function then uses the blocks, original_first_rows, and flow with `html_2_pandoc`
to produce the desired output file.
}

