% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_split_results_tables}
\alias{make_split_results_tables}
\title{Generate Results Tables Reports Split (or Grouped By) their entries in a Response Column}
\usage{
make_split_results_tables(qsf_path, csv_path, output_dir, split_by,
  headerrows = 3)
}
\arguments{
\item{qsf_path}{(optional) is the string path location of the .qsf file to be processed.}

\item{csv_path}{(optional) is the string path location of the .csv file to be processed.}

\item{output_dir}{specifies the path of the directory to save the output file in.}

\item{split_by}{is a list which specifies which columns should be used to split the respondents.}

\item{headerrows}{(optional) specifies the number of header rows in the CSV data.}
}
\description{
The make_split_results_table function works by constructing and inserting an additional
column into the responses data frame from which the responses are split. Once the responses
are split, they are inserted into distinct lists of blocks (one list for each split group of
responses) and then results tables reports are rendered from these split blocks. The function
renders these reports by looping over the list of split blocks, naming each report according to
its split respondent group, and saving each file to the specified output_dir. At the simplest
level, this function is about running get_setup, create_merged_response_column,
split_respondents, and html_2_pandoc in the right way to produce split reports.
}
