% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{mc_multiple_answer_results}
\alias{mc_multiple_answer_results}
\title{Create the Results Table for a Multiple Choice Multiple Answer Question}
\usage{
mc_multiple_answer_results(question, original_first_rows, sort_by = "N")
}
\arguments{
\item{question}{This is a list object representing an individual question
from a Qualtrics Survey File. The question must have a paired
response column placed into the question
under [['Responses']]. The insertion of the responses into questions is
handled by link_responses_to_questions.}

\item{original_first_rows}{A dataframe contianing the header information
for each column of response data. This dataframe includes a row for the DataExportTag based
response column names, another for the Question Text stem and choice text (although
truncated), and a row with QID based column names.}

\item{sort_by}{This determines how the data table is sorted. It is automatically set to
sort by N then by Choices alphebetically, but if you want to only sort by Choices alphebetically,
simply set sort_by = "choices_alpha". If you would rather the table be sorted by choice order, set sort_by = "choices_order.}
}
\description{
The mc_multiple_answer_results function uses the definition of the choices in the QSF file
and their (if present) recoded values to determine how to table the results paired to that question.
}
