% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{merge_split_column_into_comment_sheet}
\alias{merge_split_column_into_comment_sheet}
\title{Merge a Splitting Column into an Unprocessed Coded Comment Sheet}
\usage{
merge_split_column_into_comment_sheet(coded_comment_sheet, responses,
  split_column)
}
\arguments{
\item{coded_comment_sheet}{A single dataframe, imported from a
file in the format as specified by the wiki.
https://github.com/ctesta01/QualtricsTools/wiki/Comment-Coding}

\item{responses}{A dataframe of Qualtrics responses to a survey.
Use ask_for_csv() to create such a dataframe from a CSV file.}

\item{split_column}{The string name of the column to merge in for splitting}
}
\value{
A dataframe similar to the input coded_comment_sheet, but with
the additional column specified by split_column merged in.
}
\description{
Run create_merged_response_column() before this to
create the splitting column in the responses.
Then pass the column name of the column created by
create_merged_response_column() as split_column
}
