% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_get_reorganized_questions_blocks.R
\name{note_text_from_survey}
\alias{note_text_from_survey}
\title{Generate a named list of user notes from the survey. Each list element is named with the
parent question export tag and contains the survey note text prepended with "User note: ".}
\usage{
note_text_from_survey(survey)
}
\arguments{
\item{survey}{A qualtrics survey list object,
uploaded from a Qualtrics Survey File (QSF). Use
ask_for_qsf() to create such a survey list object from a QSF file.}
}
\value{
This returns a list of user notes named with the parent question export tag.
}
\description{
Generate a named list of user notes from the survey. Each list element is named with the
parent question export tag and contains the survey note text prepended with "User note: ".
}
