% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{number_of_blocks}
\alias{number_of_blocks}
\title{Count the Number of Blocks}
\usage{
number_of_blocks(blocks)
}
\arguments{
\item{blocks}{A list of blocks}
}
\value{
length of blocks; includes added elements, e.g. header
}
\description{
Previously this was used to count the number of valid question blocks.
It relied on having real question blocks indexed by R instead of named.
This function has now been updated to just return the length of block elements.
In the future, it may be updated to return the number of VALID blocks as needed.
}
