% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{percent0}
\alias{percent0}
\title{Format Percents}
\usage{
percent0(x, digits = 1, format = "f", ...)
}
\description{
percent0 is a function for formatting a decimal number into a string representing a percentage with
a specific number of decimal places and with a "%" appended.
The function uses formatC from the R base package and
defaults to formatting the number with 1 digit and with the "f" format
in the formatC argument.
}

