% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{process_question_results}
\alias{process_question_results}
\title{Append the Response Frequency Table to a Question}
\usage{
process_question_results(question, original_first_rows)
}
\description{
This function uses the contents of a question to
determine which kind of question the given question is.
Then it passes the question and the original_first_rows
to the function which processes questions of that specific
question type. After adding the results `Table` to the question,
the question is returned including this table as an additional new
list element.
}

