% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{qtype_human}
\alias{qtype_human}
\title{Determine the question type from the QSF}
\usage{
qtype_human(question)
}
\arguments{
\item{question}{A survey question}
}
\value{
Human readable question type to assist with understanding how the quuestion is
interpreted by QualtricsTools and how results are processed. Question types that are
currently unspecified in this function with return value "Other question type."
}
\description{
Determine the question type from the question's Payload information and return this as human readable text.
}
\details{
Determine the type of question based on user understanding and how the question results will be processed.
This is in the QualtricsTools intially with questions, then again when splitting
side-by-side questions into their components. The list of question types has been defined
by Tufts OIR based on their reporting needs and will be expanded as needed with additional
types of questions. Current options include Multiple Answer, Single Answer, Rank Order,
Text Entry, Side-by-side, Descriptive Box, Other question type.
}
