% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{question_description}
\alias{question_description}
\title{Create Question Description and Results Table Entry}
\usage{
question_description(question)
}
\arguments{
\item{question}{A qualtrics survey question}
}
\value{
A list of HTML with a description table for the question,
and either the generated results, or a note saying why there are
no results.
}
\description{
Create Question Description and Results Table Entry
}

