% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{questions_from_blocks}
\alias{questions_from_blocks}
\title{Generate a List of Questions from Those Contained in the Blocks}
\usage{
questions_from_blocks(blocks)
}
\arguments{
\item{blocks}{A list of the survey blocks, with the questions and responses inserted
in place of the BlockElements. Structuring the blocks in this way is automatically
handled by get_reorganized_questions_and_blocks, but can also be performed by (after
inserting the responses into questions) using the questions_into_blocks method.}
}
\description{
This function iterates through the blocks and anything that has a DataExportTag
is added to a list of questions, and it returns that list of questions from
the blocks.
}
