% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{questions_from_survey}
\alias{questions_from_survey}
\title{Generate a List of Questions}
\usage{
questions_from_survey(survey)
}
\value{
A list of questions from the uploaded QSF file
}
\description{
This function takes the questions out of the survey and
turns them into their own list for later use.
Each question is an element of the returned list, and
each element has its own list as its content. Most questions
include things like "SurveyID", "Element", "PrimaryAttribute",
"SecondaryAttribute", and a "Payload" which contains
most of the information about the question.
}

