% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{questions_into_blocks}
\alias{questions_into_blocks}
\title{Organize the Questions into their Survey Blocks}
\usage{
questions_into_blocks(questions, blocks)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics survey}

\item{blocks}{A list of blocks from a Qualtrics survey}
}
\value{
a list of blocks including questions under blocks[[i]]$BlockElements
for each index i.
}
\description{
This function takes a blocks list (create this with blocks_from_survey and
then remove_trash_blocks), inserts the questions from a survey
appropriately into the blocks element, and then returns the
blocks list including the questions as elements in the blocks[[i]]$BlockElements.
}

