% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{remove_trash_blocks}
\alias{remove_trash_blocks}
\title{Remove the Trash Block from the list of Blocks}
\usage{
remove_trash_blocks(blocks)
}
\arguments{
\item{blocks}{A list of blocks from a Qualtrics survey}
}
\value{
The list of blocks is returned without any Trash blocks
}
\description{
This function finds among the trash blocks which has its [['Type']] value
set to "Trash" and then removes it from the blocks list. The iteration
in this function is backwards because it assigns NULL to any list
items which need to be removed. Therefore, if it assigns NULL
to a value and then moves up, it will not only skip questions as the
higher up questions move downward as questions are deleted, but
the total length of the list will be changing as it gets closer to the
end of the list.
}

