% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{remove_trash_questions}
\alias{remove_trash_questions}
\title{Remove Questions from the Trash Block}
\usage{
remove_trash_questions(questions, blocks)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics survey}

\item{blocks}{A list of blocks from a Qualtrics survey}
}
\value{
The list of questions returned is the list of questions
provided except without any questions listed in the Trash
block of the blocks list provided.
}
\description{
This function removes any questions from a list of questions
that are listed in the Trash block.
}

