% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_responses}
\alias{sample_responses}
\title{Sample Qualtrics Response Data}
\format{An object of class \code{data.frame} with 10 rows and 38 columns.}
\usage{
sample_responses
}
\description{
This is some sample response data, to the matching sample survey provided
}
\keyword{datasets}
