% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_survey}
\alias{sample_survey}
\title{Sample Qualtrics Survey Data}
\format{An object of class \code{list} of length 2.}
\usage{
sample_survey
}
\description{
This is a sample survey, with a couple questions of different types.
}
\keyword{datasets}

