% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{split_sbs_questions}
\alias{split_sbs_questions}
\title{Split side by side question into their component question parts}
\usage{
split_sbs_questions(questions)
}
\arguments{
\item{questions}{A list of questions pulled from Survey Elements of a Qualtrics QSF}
}
\value{
A list of questions with side-by-side questions split into their component question parts.
}
\description{
Split side by side question into their component question parts
}
