% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{split_sbs_questions_blocks}
\alias{split_sbs_questions_blocks}
\title{Split side by side questions and corresponding block elements}
\usage{
split_sbs_questions_blocks(questions, blocks)
}
\arguments{
\item{questions}{A list of Qualtrics survey questions that may include side-by-side.}

\item{blocks}{A list of Qualtrics blocks for which a side-by-side question is a single block element.}
}
\value{
A named list with two elements: "questions" is a list of survey questions
with side-by-side questions split into their component questions, and "blocks" is a list
of survey blocks within which side-by-side question block elements have been replaced with
spaces for each of the component question elements.
}
\description{
Split side-by-side survey questions into their component questions. Create
spaces within the blocks for these split elements.
}
