% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{split_side_by_side_q}
\alias{split_side_by_side_q}
\title{Split Side-by-Side Questions into Multiple Component Questions}
\usage{
split_side_by_side_q(question)
}
\arguments{
\item{question}{A question with additional question elements}
}
\value{
A list of with each component question split out of the side-by-side
}
\description{
Given a side-by-side question, split this into separate questions list elements for each
of the multiple question parts.
}
\details{
This function is called within the QualtricsTools setup process and splits
multiple questions into separate component questions so that results may be processsed.
QualtricsTools is designed to process results to side-by-side questions a separate questions.
When this is called as part of the setup process, there is an earlier check to
ensure that questions passed to this function are actual side-by-side questions.
}
