% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{split_side_by_sides}
\alias{split_side_by_sides}
\title{Split Side-by-Side Questions into Multiple Questions}
\usage{
split_side_by_sides(questions, blocks)
}
\arguments{
\item{questions}{A list of questions from a survey}

\item{blocks}{A list of blocks from a survey}
}
\value{
A list of questions and a list of blocks with their SBS questions split
into multiple questions
}
\description{
This function updates both the list of questions and list of blocks from a survey
to reflect a side-by-side question as multiple individual questions.
}
