% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{tabelize_display_logic}
\alias{tabelize_display_logic}
\title{Generate Tables for Each Question with Display Logic}
\usage{
tabelize_display_logic(blocks)
}
\arguments{
\item{blocks}{Survey blocks with questions substituted for the blockelements}
}
\value{
a list of html tables detailing the display logic for each question
containing display logic.
}
\description{
This function loops through every block, and checks if
each block element has display logic (and is not a descriptive box).
The descriptive boxes aren't included.
If it has display logic, it's inserted into the tables returned by
this function.
}
