% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{table_html_coded_comments}
\alias{table_html_coded_comments}
\title{Create a Text Appendix for Coded Comments}
\usage{
table_html_coded_comments(question, cc_index, appendix_e, blocks,
  original_first_row)
}
\arguments{
\item{question}{is the question to which the coded comments belong.}

\item{cc_index}{is the index of the coded comments to appendicize.}

\item{appendix_e}{is the number of the appendix in the text appendices report.}

\item{blocks}{is the list of survey blocks.}

\item{original_first_row}{is the original header information from the CSV response data.}
}
\description{
This returns a list containing two tables: the question's description table and the categorical
breakdown of the responses into the coded comments' categories. The description includes an
"Appendix ##" title, the question text, and a tag which states the following appendix is
"Coded Comments." The coded comments table has two columns: the "Responses" and "N", in which
the categories and their frequencies are listed.
}

