% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{table_mcsa_multitext}
\alias{table_mcsa_multitext}
\title{Create a message stating MCSA questions with Multiple Text Entry components
can't be automatically processed for text appendices.}
\usage{
table_mcsa_multitext(question)
}
\description{
Create a message stating MCSA questions with Multiple Text Entry components
can't be automatically processed for text appendices.
}
