% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{table_text_entry}
\alias{table_text_entry}
\title{Create Text Appendices for Text Entry Type Questions}
\usage{
table_text_entry(question, responses, appendix_e, blocks, original_first_row)
}
\arguments{
\item{question}{is the question for which the text appendix is being created.}

\item{responses}{is a data frame of text responses to the question given.}

\item{appendix_e}{is the number of the appendix in the text appendices report.}

\item{blocks}{is the list of survey blocks.}

\item{original_first_row}{is the original header information from the CSV response data.}
}
\description{
The `table_text_entry` function creates a text appendix with as many
columns as there are text entry components to the given question and fills
them with the responses from each respondent across the rows.
}
