% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{uncodeable_question_dictionary}
\alias{uncodeable_question_dictionary}
\title{Create Uncodeable Question Dictionary}
\usage{
uncodeable_question_dictionary(blocks)
}
\description{
Create Uncodeable Question Dictionary
}

