% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{valid_questions_blocks_from_survey}
\alias{valid_questions_blocks_from_survey}
\title{Pull the valid questions and blocks from a Qualtrics Survey.}
\usage{
valid_questions_blocks_from_survey(survey)
}
\arguments{
\item{survey}{A qualtrics survey list object,
uploaded from a Qualtrics Survey File (QSF). Use
ask_for_qsf() to create such a survey list object from a QSF file.}
}
\value{
List with named elements questions and blocks. this includes only valid
questions. Trash questions have been removed from both questions and blocks.
Questions are named by their data export tags, blocks named by their descriptive names.
Block elements for questions are name by the QIDs.
}
\description{
This pulls the questions and blocks from a QSF survey file.
Trash questions are removed from the blocks and the list of questions.
Side-by-side questions are split into their component question parts for ease
of processing results.
}
