% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{validate_data_export_tags}
\alias{validate_data_export_tags}
\title{Validate Data Export Tag Uniqueness}
\usage{
validate_data_export_tags(questions)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics survey}
}
\description{
It is crucial for parts of this program to work that the DataExportTags be unique
to one another. This program checks for duplicates among the DataExportTags, and
returns a boolean statement representing whether or not the DataExportTags validated.
validate_export_tags returns TRUE if the DataExportTags are not duplicated, and are valid,
and FALSE if the DataExportTags did not validate, and are duplicated.
}

