% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{validate_response_columns}
\alias{validate_response_columns}
\title{Validate Response Column Uniqueness}
\usage{
validate_response_columns(responses)
}
\arguments{
\item{responses}{A data frame of responses from a Qualtrics survey}
}
\description{
It is crucial for parts of this program to work that the data response columns be unique
to one another. This program checks for duplicates among the response column names
of the questions and returns FALSE if there are any duplicates, and TRUE if there are no
duplicates. Equivalently, FALSE represents FALSE validation, and TRUE represents TRUE validation.
}

