% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{which_choices_have_text_entry}
\alias{which_choices_have_text_entry}
\title{Find out Which Choices Have Text Entry Components}
\usage{
which_choices_have_text_entry(question)
}
\arguments{
\item{question}{This is a list object representing an individual question
from a Qualtrics Survey File. A question must have the 'Choices' list
inside its 'Payload' in order to be accepted by this function.}
}
\value{
A list of the indices for the choices which have a text entry component.
}
\description{
Given a single question, this function
determines which of the question[['Payload']][['Choices']]
elements have a TextEntry which is enabled for free response
input. The returned value is a list of integers representing
the indices of which choices have text entry components.
}
