% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{appendix_lettering}
\alias{appendix_lettering}
\title{`appendix_lettering` takes a number
and returns the corresponding lettered index.
examples:
1 -> A
2 -> B
27 -> AA
29 -> AC
and so on.}
\usage{
appendix_lettering(number)
}
\description{
`appendix_lettering` takes a number
and returns the corresponding lettered index.
examples:
1 -> A
2 -> B
27 -> AA
29 -> AC
and so on.
}

