% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_files.R
\name{ask_user_for_qsf}
\alias{ask_user_for_qsf}
\title{Ask the user for the Qualtrics Survey file}
\usage{
ask_user_for_qsf()
}
\value{
The survey file the user uploads, as a list
}
\description{
Ask the user for the Qualtrics Survey file
}

