% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{choice_text_from_question}
\alias{choice_text_from_question}
\title{Get the Choice Text based on the Choice from a Question}
\usage{
choice_text_from_question(question, choice)
}
\description{
Input a question and a choice, and this function will
return the choice text.
}

