% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{clean_html_and_css}
\alias{clean_html_and_css}
\title{Clean HTML and CSS from a string}
\usage{
clean_html_and_css(text)
}
\arguments{
\item{text}{any text string that might contain HTML or whitespace that needs to be
stripped.}
}
\value{
text without any html, css or extraneous whitespace, and with breaks replaced by space character.
}
\description{
This function uses regex extensively to clean HTML and CSS out of a given text block.
}
\details{
This function uses regular expressions to clean HTML and CSS from text. This is used
in various places to clean quesiton and choice text.
All line breaks and non-break space are replaced with a space character.
HTML tags are removed, so HTML formatting will NOT be included in any reports.
This also removes CSS based on patterns specific to matrix formatting and row banding
used by Tufts OIR.
At the end, all leading, trailing and extra whitespace are removed, and repeated
whiteshapce is replaced by a single whitespace character.
}
