% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{create_question_dictionary}
\alias{create_question_dictionary}
\title{Create a Question Dictionary}
\usage{
create_question_dictionary(blocks)
}
\description{
Create a Question Dictionary
}

