% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{create_question_dictionary}
\alias{create_question_dictionary}
\title{Create a Question Dictionary}
\usage{
create_question_dictionary(blocks)
}
\arguments{
\item{blocks}{The blocks provided to this function must include questions inserted into
the BlockElements. Create the list of blocks from a survey with blocks_from_survey(),
and with questions on hand, insert them into the blocks with questions_into_blocks().}
}
\value{
A data frame with a row for each question describing the question's details.
}
\description{
Create a Question Dictionary
}

