% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{directory_get_coded_comment_sheets}
\alias{directory_get_coded_comment_sheets}
\title{Turn a Directory into a list of Coded Comment Data Frames (unprocessed)}
\usage{
directory_get_coded_comment_sheets(directory)
}
\description{
Turn a Directory into a list of Coded Comment Data Frames (unprocessed)
}

