% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{directory_get_coded_comment_sheets}
\alias{directory_get_coded_comment_sheets}
\title{Turn a Directory into a list of Coded Comment Data Frames (unprocessed)}
\usage{
directory_get_coded_comment_sheets(directory)
}
\description{
This function takes as an argument a string representative of a
directory, loads the CSVs and Excel data from that directory,
looks for 'Coded' sheets, extracts the coded sheets,
and saves the coded comment table and the question ID
as a pair in the output coded_appendix_tables list. If there
are sheets which contain non-numeric data, warnings are raised.
}

