% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_question_index}
\alias{find_question_index}
\title{Find Question Index from DataExportTag}
\usage{
find_question_index(questions, exporttag)
}
\description{
This function takes a list of questions and an export tag and
looks for the matching question. It returns the index(es) of
the questions with that Question Data Export Tag.
}
