% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{format_and_split_comment_sheets}
\alias{format_and_split_comment_sheets}
\title{Format and Split a list of Unprocessed Coded Comment Sheets}
\usage{
format_and_split_comment_sheets(coded_comment_sheets, responses, split_column)
}
\arguments{
\item{split_column}{The string name of the column across which the coded comments
should be split.}
}
\value{
A list of lists, for each split group, and for the pairs of question IDs with
coded comment tables in each split group.
}
\description{
When splitting the respondents of a survey to create split reports,
the coded comments are split by this function and then returned as a list of lists.
The first list is a list for each split group, and each list within those is a
list of pairs of question IDs and their coded comments tables.
}
