% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{format_coded_comments}
\alias{format_coded_comments}
\title{Turn the original coded comments sheet into a pair: (Question, Data Frame)}
\usage{
format_coded_comments(coded_comment_sheet)
}
\description{
Turn the original coded comments sheet into a pair: (Question, Data Frame)
}

