% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{generate_summary_stats}
\alias{generate_summary_stats}
\title{Create Summary Statistics tables for numerical data only questions}
\usage{
generate_summary_stats(question, horizontal_table = FALSE)
}
\arguments{
\item{question}{This is the question which has been checked to contain
Text Entry data with Numerical verification. This TE question will now
be processed to give a summary stats table and not an appendice.}

\item{horizontal_table}{This variable indicated whether the summary statistics
table should be presented with each statistic in a separate column. By default it
is set to FALSE.}
}
\value{
A question with a ['Table'] appended to it where the summary stats are
present
}
\description{
This function uses the data frame of responses that
have been linked to a question to generate summary statistics using
basic R functions mean(), median(), min(), max(), sd(). It creates a
new data frame with colnames and the stats then appends it to question['Table'].
}
