% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{get_coded_comment_sheet}
\alias{get_coded_comment_sheet}
\title{Turn a Single Coded File into a Data Frame}
\usage{
get_coded_comment_sheet(codedfile)
}
\arguments{
\item{codedfile}{The string path to a Excel or CSV file.}
}
\value{
A dataframe version of the contents of the coded comments
in the codedfile.
}
\description{
This retrieves comment coding data as a dataframe
from a Excel or CSV file. If the file is an Excel document,
the coded comments are retrieved from the sheet named "Coded"
(case does not matter), and if the file is a CSV it is directly
read as a dataframe.
}
