% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{get_coded_comment_sheet}
\alias{get_coded_comment_sheet}
\title{Turn a Single Coded File into a Data Frame}
\usage{
get_coded_comment_sheet(codedfile)
}
\description{
Turn a Single Coded File into a Data Frame
}

