% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{get_coded_comment_sheet_NVivo}
\alias{get_coded_comment_sheet_NVivo}
\title{Turn a Single NVivo Crosstab Coded File into a Data Frame}
\usage{
get_coded_comment_sheet_NVivo(codedfile)
}
\arguments{
\item{codedfile}{The string path to a Excel file.}
}
\value{
A dataframe version of the contents of the coded comments
in the codedfile.
}
\description{
This retreives comment coding data as a data frame. Data must be
formatted with NVivo crosstab export format with ResponseID unlabeled
in the first column and second column labeled with the variable name
and values indicating response presence. Subsequent columns indicate
coded comment categories.
}
