% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{get_setup}
\alias{get_setup}
\title{Setup the Global Environment for a Survey}
\usage{
get_setup(headerrows)
}
\arguments{
\item{headerrows}{An optional parameter for specifying the number of
headerrows in the response csv.}
}
\description{
This function sets up the .GlobalEnv to include survey, responses, questions
(without trash questions, with clean question text,
with a human readable question type, with responses, and with results if
automatically generateable), and blocks (without trash blocks, with questions
inserted in place of the BlockElements representing them).
}

