% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{insert_coded_comments}
\alias{insert_coded_comments}
\title{Insert Coded Comments into Blocks}
\usage{
insert_coded_comments(blocks, original_first_rows, coded_comments)
}
\arguments{
\item{blocks}{A list of the survey blocks, with the questions included in them.}

\item{original_first_rows}{A dataframe contianing the header information
for each column of response data. This dataframe should include a row for the DataExportTag based
response column names, another for the Question Text stem and choice text (although
truncated), and a row with QID based column names.}

\item{coded_comments}{A list of pairs (varname, coded_table) where varname corresponds
to the response column name of the comments coded and coded_table
summarizes the frequencies of the provided coded comments.}
}
\description{
This takes a list of pairs of question IDs and coded comments tables
and finds their corresponding question based on the question ID in a list
of blocks and then inserts the coded comments table into the question.
The returned list is a list of blocks where the questions have had their
coded comments inserted.
}
