% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{insert_coded_comments}
\alias{insert_coded_comments}
\title{Insert Coded Comments into Blocks}
\usage{
insert_coded_comments(blocks, original_first_rows, coded_comments)
}
\description{
This takes a list of pairs of question IDs and coded comments tables
and finds their corresponding question based on the question ID in a list
of blocks and then inserts the coded comments table into the question.
The returned list is a list of blocks where the questions have had their
coded comments inserted.
}

