% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{insert_questions_into_block}
\alias{insert_questions_into_block}
\title{Insert questions into blocks}
\usage{
insert_questions_into_block(questions, block)
}
\arguments{
\item{questions}{A list of questions from a Qualtrics survey}

\item{block}{A single block from a Qualtrics survey}
}
\value{
The single survey block with question elements updated to include the full question. In the setup process,
BlockElements within the blocks that correspond to survey questions are initially named with question QIDs. This function
renames the Block Elements with question Data Export Tags corresponding to the question list.
}
\description{
This inserts questions into a single block. It should be iterated over all blocks
to fully update the survey.
}
