% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{insert_split_survey_comments}
\alias{insert_split_survey_comments}
\title{Split Survey and Insert Split Coded Comments}
\usage{
insert_split_survey_comments(split_blocks, split_coded_comment_sheets,
  split_column, original_first_rows)
}
\description{
The responses should already include the split column
}

