% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_coding.R
\name{insert_split_survey_comments}
\alias{insert_split_survey_comments}
\title{Split Survey and Insert Split Coded Comments}
\usage{
insert_split_survey_comments(split_blocks, split_coded_comment_sheets,
  split_column, original_first_rows)
}
\arguments{
\item{split_blocks}{A list of a list of blocks. The same question, but with different respondent groups,
might look something like split_blocks[[1]][[1]][['BlockElements']][[1]] and
split_blocks[[2]][[1]][['BlockElements']][[1]]. These refer to the first and second respondent
groups, the first block, and the first block element. Split blocks of this form are
generated by the split_respondents function.}

\item{split_coded_comment_sheets}{A list of lists where the outer lists each split group
and the inner lists the pairs of question IDs with coded comment tables in each split group.}

\item{split_column}{The string name of the column across which the coded comments
should be split.}

\item{original_first_rows}{A dataframe contianing the header information
for each column of response data. This dataframe should include a row for the DataExportTag based
response column names, another for the Question Text stem and choice text (although
truncated), and a row with QID based column names.}
}
\description{
This function splits the survey's response data and the
coded comments, then inserts the split coded comments into the split
surveys.
}
