% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_matrix_question}
\alias{is_matrix_question}
\title{Determine if a question is a matrix question}
\usage{
is_matrix_question(x)
}
\description{
Determine if a question is a matrix question
}

