% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_type_checking.R
\name{is_multiple_choice}
\alias{is_multiple_choice}
\title{Determine if a question is a multiple choice question}
\usage{
is_multiple_choice(x)
}
\description{
Determine if a question is a multiple choice question
}
