% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{lean_responses}
\alias{lean_responses}
\title{Create Long and Lean Response Dictionary}
\usage{
lean_responses(question_blocks, survey_responses)
}
\arguments{
\item{question_blocks}{A list of blocks, with questions inserted in place of the
BlockElements representing them.}

\item{survey_responses}{The responses to the survey, as imported by ask_for_csv()}
}
\value{
a data frame with each row detailing an individual survey response.
}
\description{
lean_responses() creates a data frame where each row corresponds to
an individual response to the survey. Each response contains
the respondents' id, the response column's name,
the variable response, and the coded
response.
}
