% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{matrix_multiple_answer_results}
\alias{matrix_multiple_answer_results}
\title{Create the Results Table for a Matrix Multiple Answer Question}
\usage{
matrix_multiple_answer_results(question)
}
\arguments{
\item{question}{The question must have a paired response column placed into the question
under $Responses}
}
\value{
a table with the matrix-sub-questions listed in the first column,
a column with the total number of respondents for each subquestion, and
the percentages for each choice for each sub-question.
}
\description{
Create the Results Table for a Matrix Multiple Answer Question
}

