% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{matrix_single_answer_results}
\alias{matrix_single_answer_results}
\title{Create the Results Table for a Matrix Single Answer Question}
\usage{
matrix_single_answer_results(question)
}
\arguments{
\item{question}{The question must have a paired response column placed into the question
under `$Responses`}
}
\value{
a table with the matrix-sub-questions listed in the first column, the percentages for each
choice for each sub-question listed in a table, and then another column with the total respondents
for each subquestion.
}
\description{
The matrix_single_answer_results function uses the definition of the choices and answers in the
QSF file and their potentially recoded values to determine how to table the results paired
to that question. If you look at the source code, keep in mind that a matrix question's sub-questions
are called "Choices" and that the choices for each sub-question are called "Answers"
}

