% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{mc_single_answer_results}
\alias{mc_single_answer_results}
\title{Create the Results Table for a Multiple Choice Single Answer Question}
\usage{
mc_single_answer_results(question, original_first_rows)
}
\arguments{
\item{question}{The question must have a paired response column placed into the question
under [['Responses']]}
}
\value{
a table with an N, Percent, and choice column, detailing the number of responses for each
choice.
}
\description{
The mc_single_answer_results function uses the definition of the choices in the QSF file
and their potentially recoded values to determine how to table the results paired to that question.
}

