% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{notes_from_survey}
\alias{notes_from_survey}
\title{Generate a List of Notes Blocks}
\usage{
notes_from_survey(survey)
}
\arguments{
\item{survey}{This should be a Qualtrics survey in the form of a list
imported from the JSON-formatted QSF file.}
}
\value{
This returns a list of blocks
}
\description{
Generate a List of Notes Blocks
}

