% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{notes_from_survey}
\alias{notes_from_survey}
\title{Generate a List of Notes Blocks}
\usage{
notes_from_survey(survey)
}
\arguments{
\item{survey}{A qualtrics survey list object,
uploaded from a Qualtrics Survey File (QSF). Use
ask_for_qsf() to create such a survey list object from a QSF file.}
}
\value{
This returns a list of blocks
}
\description{
Generate a List of Notes Blocks
}
