% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{number_of_blocks}
\alias{number_of_blocks}
\title{Count the Number of Blocks}
\usage{
number_of_blocks(blocks)
}
\arguments{
\item{blocks}{A list of blocks}
}
\value{
the number of question blocks
}
\description{
Since the blocks list is used to transport some additional information
beyond the contents of the survey questions, this function is here to
help in counting how many valid question blocks there are.
Any real question blocks will be enumerated (aka numbered) in R, as opposed to the
content that's been added which will be named (with a string). This means that when
looking at the names of the blocks list, the integer values or the values which
have no name are the question blocks, and the values which have names are the
information added by the QualtricsTools functions. This function counts up the former.
}
