% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{question_from_response_column}
\alias{question_from_response_column}
\title{Get the Index of the Question Corresponding to a Response Column}
\usage{
question_from_response_column(blocks, response_name)
}
\description{
Use this function to get the indexes identifying a question
in the blocks list. Give it a response column name, and it will
try to find the question it corresponds to. Otherwise, it will
respond NULL.
}
