% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{remove_trash_blocks}
\alias{remove_trash_blocks}
\title{Remove the Trash Block from the list of Blocks}
\usage{
remove_trash_blocks(blocks)
}
\arguments{
\item{blocks}{A list of blocks from a Qualtrics survey}
}
\value{
The list of blocks is returned without any Trash blocks
}
\description{
This function finds among the trash blocks which has its $Type value
set to "Trash" and then removes it from the blocks list.
}

