% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizing_survey_data.R
\name{split_respondents}
\alias{split_respondents}
\title{Split Respondents by a Response Column}
\usage{
split_respondents(response_column, responses, survey, blocks, questions,
  headerrows, already_loaded, original_first_rows)
}
\description{
This function splits the respondents into separate respondent groups
based on the values in the specified response column. Then, for each
respondent group, the blocks with questions are duplicated. Each set of
blocks has a different responent group inserted into its questions, and
then each set of blocks is processed. The output is a list of blocks with
the results processed and inserted into each BlockElement.
}
\details{
@param response_column The response column that will be used to split the respondents
 @param headerrows the number of rows in the response csv before the response data starts
 @param already_loaded This can be set to TRUE to indicate that the survey and responses
 should be sourced from the global scope; in other words that the survey and its responses
 have are "already loaded."

@return A list of a list of blocks. The same question, but with different respondent groups,
 might look something like split_blocks[[1]][[1]][['BlockElements']][[1]] and
 split_blocks[[2]][[1]][['BlockElements']][[1]]. These refer to the first and second respondent
 groups, the first block, and the first block element.
}
