% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_generation.R
\name{tabelize_blocks}
\alias{tabelize_blocks}
\title{Create a List of HTML Versions of the Results Tables}
\usage{
tabelize_blocks(blocks)
}
\arguments{
\item{questions}{A list of questions with the relevant results tables
stored as data frames under the questions[[i]]$Table element. Create
such a list of questions by using generate_results function}
}
\value{
A list of HTML results tables for each question
}
\description{
Create a List of HTML Versions of the Results Tables
}

