% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{table_no_respondents}
\alias{table_no_respondents}
\title{Create a standard "No Respondents answered this question" text appendix}
\usage{
table_no_respondents(question, appendix_e)
}
\arguments{
\item{question}{is the question for which the text appendix is being created.}

\item{appendix_e}{is the number of the appendix in the text appendices report.}
}
\description{
This function returns a standardized html table intended to go into a
document of text appendices indicating that there were no respondents
who answered the given text entry question.
}
