% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{table_non_text_entry}
\alias{table_non_text_entry}
\title{Create Text Appendices for Non-Text Entry Type Questions}
\usage{
table_non_text_entry(question, responses, appendix_e, blocks,
  original_first_row)
}
\arguments{
\item{question}{is the question for which the text appendix is being created.}

\item{responses}{is a data frame of text responses to the question given.}

\item{appendix_e}{is the number of the appendix in the text appendices report.}

\item{blocks}{A list of blocks with block elements replaced
by the question with its paired responses.}

\item{original_first_row}{This is the original first row of the
response set. If you have the original_first_rows, you can
pass original_first_rows[1,] to}
}
\description{
This function creates text appendices, with a description table and
responses table, for questions which are not text-entry type questions.
The description table contains a text appendix title, like "Appendix A"
and the responses are tabled verbatim beneath it.
}
