% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{text_appendices_table}
\alias{text_appendices_table}
\title{Create HTML Tables for the Text Entry Questions}
\usage{
text_appendices_table(blocks, original_first_row, flow)
}
\arguments{
\item{blocks}{A list of blocks with block elements replaced
by the question with its paired responses.}

\item{original_first_row}{This is the original first row of the
response set. If you have the original_first_rows, you can
pass original_first_rows[1,] to}
}
\value{
an html string containing a title,
question text, and the text responses for each
text appendix.
}
\description{
This function creates an HTML string with tables
for each of the text entry questions and their responses.
The appendix_lettering function inside this function
is to create lettering for the tables in the style
of "A, B, ..., Z, AA, ..., ZZ" and so forth. The html
tables are created by looping through the blocks
and their contained question block elements. For each
question, if the question is a Text Entry question or
any of the response columns contain "TEXT" then a table
is created for those responses.
}

