% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_results.R
\name{uncodeable_questions_message}
\alias{uncodeable_questions_message}
\title{Create a Message Stating Which Questions Weren't Automatically Tabled}
\usage{
uncodeable_questions_message(questions)
}
\value{
A message stating for which questions could not have
results automatically generated.
}
\description{
This is function is used in the Shiny app to tell users which questions weren't
automatically coded. This may be changed later to be more informative, or
to include this information elsewhere.
}
